% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twolines.R
\name{eval2}
\alias{eval2}
\title{Evaluate a String as an Expression}
\usage{
eval2(string)
}
\arguments{
\item{string}{A character string containing an R expression to evaluate.}
}
\value{
The result of evaluating the expression.
}
\description{
Internal helper function that evaluates a string as an R expression.
Used to dynamically access variables by name.
}
\keyword{internal}
