% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_freq.R
\name{fhist}
\alias{fhist}
\title{Deprecated alias for plot_freq}
\usage{
fhist(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link{plot_freq}}}
}
\description{
This function is deprecated. Use \code{\link{plot_freq}} instead.
}
