% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get.colors}
\alias{get.colors}
\title{Get default colors for plotting}
\usage{
get.colors(k)
}
\arguments{
\item{k}{Number of colors to return}
}
\value{
A character vector of color names
}
\description{
Returns a vector of colors based on the number of groups.
This is an internal utility function used by plot_cdf() and plot_density().
}
\details{
Color schemes:
\itemize{
  \item 2 groups: red4, dodgerblue
  \item 3 groups: red4, dodgerblue, green4
  \item 4 groups: orange1, orange3, red2, red4
  \item 5+ groups: extends the 4-group palette with additional colors (dodgerblue, blue, green, purple, gray)
}
}
