% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm2.R
\name{predict.lm2}
\alias{predict.lm2}
\title{Predict method for lm2 objects}
\usage{
\method{predict}{lm2}(object, ...)
}
\arguments{
\item{object}{An object of class \code{lm2}}

\item{...}{Additional arguments passed to \code{\link[estimatr]{predict.lm_robust}},
including \code{newdata}, \code{se.fit}, and \code{interval}.}
}
\value{
A vector of predicted values (or a data frame if \code{se.fit = TRUE} or 
  \code{interval} is specified)
}
\description{
Predict method for lm2 objects
}
