% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc_var.R
\name{print.desc_var}
\alias{print.desc_var}
\title{Print method for desc_var objects}
\usage{
\method{print}{desc_var}(x, ...)
}
\arguments{
\item{x}{An object of class \code{desc_var}}

\item{...}{Additional arguments passed to print.data.frame}
}
\value{
Invisibly returns the original object
}
\description{
Print method for desc_var objects
}
