% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm2.R
\name{summary.lm2}
\alias{summary.lm2}
\title{Summary method for lm2 objects}
\usage{
\method{summary}{lm2}(object, ...)
}
\arguments{
\item{object}{An object of class \code{lm2}}

\item{...}{Additional arguments passed to \code{\link{print.lm2}}}
}
\value{
Invisibly returns the original object
}
\description{
Summary method for lm2 objects
}
