% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_table2}
\alias{validate_table2}
\title{Validate Inputs for table2() Function}
\usage{
validate_table2(..., data = NULL, func_name = "table2", data_name = NULL)
}
\arguments{
\item{...}{One or more variables to be tabulated.}

\item{data}{An optional data frame containing the variables.}

\item{func_name}{Character string. Name of the calling function (for error messages).
Default is "table2".}

\item{data_name}{Character string. Name of the data argument (for error messages).
If NULL, will attempt to infer from the call.}
}
\value{
A list containing:
\itemize{
  \item \code{dots}: List of evaluated variables (ready for base::table)
  \item \code{dot_expressions}: List of expressions (for variable name extraction)
  \item \code{data_name}: Name of data argument (for error messages)
}
}
\description{
Validates inputs for table2() function that accepts multiple variables via ...
with optional data frame.
}
\keyword{internal}
