% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_freq.R
\name{plot_freq}
\alias{plot_freq}
\title{Plot frequencies of a variable, optionally by group (histogram without binning)}
\usage{
plot_freq(
  x,
  freq = TRUE,
  col = "dodgerblue",
  lwd = 9,
  value.labels = TRUE,
  add = FALSE,
  by = NULL,
  ...
)
}
\arguments{
\item{x}{A numeric vector of values to plot frequencies for.}

\item{col}{Color for the bars.}

\item{lwd}{Line width for the frequency bars. Default is 9.}

\item{value.labels}{Logical. If TRUE, displays frequencies on top of each line.}

\item{add}{Logical. If TRUE, adds to an existing plot instead of creating a new one.}

\item{by}{A grouping variable (with 2 or 3 unique values). If specified, frequencies are computed separately for each group and plotted with different colors.}

\item{...}{Pass on any argument accepted by \code{plot()} e.g., \code{xlab='x-axis'} , \code{main='Distribution of X'}}
}
\value{
Invisibly returns a data frame with values and their frequencies.
}
\description{
Creates a frequency plot showing the frequency of every observed value,
displaying the full range from minimum to maximum value.
}
\details{
This function creates a frequency plot where each observed value is shown
with its frequency. Unlike a standard histogram, there is no binning, unlike
a barplot, non-observed values of the variable are shown with 0 frequency 
instead of skipped.
}
\examples{
# Simple example
x <- c(1, 1, 2, 2, 2, 5, 5)
plot_freq(x)

# Pass on some common \code{plot()} arguments
plot_freq(x, col = "steelblue", xlab = "Value", ylab = "Frequency",ylim=c(0,7))

# Add to an existing plot
plot_freq(x, col = "dodgerblue")
plot_freq(x + 1, col = "red", add = TRUE)

}
