% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_simplified_ttest.R
\name{print.simplified_ttest}
\alias{print.simplified_ttest}
\title{Print method for simplified t-test output}
\usage{
\method{print}{simplified_ttest}(x, ...)
}
\arguments{
\item{x}{An object of class \code{simplified_ttest}}

\item{...}{Additional arguments passed to print}
}
\description{
Print method for simplified t-test output
}
